Refactor tests and fix writing Shar through pipes
