+++ lhotse/shar/__init__.py
@@ -1,0 +1,2 @@
+from .readers import *
+from .writers import *
