+++ lhotse/shar/readers/__init__.py
@@ -1,1 +1,2 @@
 from .datapipes import load_shar_datapipe
+from .lazy import LazySharIterator
