+++ lhotse/shar/writers/__init__.py
@@ -1,0 +1,5 @@
+from .array import ArrayTarWriter
+from .audio import AudioTarWriter
+from .cut import CutShardWriter
+from .shar import SharWriter
+from .tar import TarWriter
